+++ CHANGELOG.md
@@ -11,7 +11,7 @@ All notable changes to this project will be documented in this file. This change
 - Bugfix - guard against nil values for required types in the execution phase
 - Add preparation and memoization as an example in executor test
 - Optionally add a `:resolver` key to state when validating a statement to inline resolver-fns prior to execution
-- Streamline validation output document
+- Streamline statement validation output - remove :state key in the absence of errors
 
 ## [0.1.19] - 2016-11-16
 - Support for Input Objects and their fields in Introspection Schema. By Boris Jonica (bjonica)